убраны аватары, добавлены сервисы и репозитории

+++ src/main/java/com/example/catsanddogstelegram/controller/DogController.java
@@ -49,7 +49,7 @@ public class DogController {
             }, tags = "Dog"
     )
     @GetMapping("{id}")
-    public ResponseEntity<Dog> read(@PathVariable(name = "id") long animalId) {
+    public ResponseEntity<Dog> read(@PathVariable(name = "id") int animalId) {
         Dog dog = dogService.read(animalId);
         return ResponseEntity.ok(dog);
     }
@@ -67,7 +67,7 @@ public class DogController {
             }, tags = "Dog"
     )
     @PutMapping("{id}")
-    public ResponseEntity<Dog> update(@PathVariable(name = "id") long animalId,
+    public ResponseEntity<Dog> update(@PathVariable(name = "id") int animalId,
                                       @RequestBody Dog dog) {
         Dog foundDog = dogService.update(animalId, dog);
         return ResponseEntity.ok(foundDog);
@@ -86,7 +86,7 @@ public class DogController {
             }, tags = "Dog"
     )
     @DeleteMapping("{id}")
-    public ResponseEntity<Dog> delete(@PathVariable(name = "id") long animalId) {
+    public ResponseEntity<Dog> delete(@PathVariable(name = "id") int animalId) {
         dogService.delete(animalId);
         return ResponseEntity.ok().build();
     }
